requirements.txt.txt
discord.py==2.4.1
Pillow==10.0.1
requests==2.31.0